Begin implementing Issuer interface for email and github identities
